+++ packages/runtime/src/widgets/Collapsible.tsx
@@ -188,14 +188,17 @@ export const Collapsible: React.FC<CollapsibleProps> = (props) => {
         },
       }}
     >
-      <Collapse
-        accordion={values?.isAccordion}
-        activeKey={activeValue}
-        expandIcon={expandIcon}
-        expandIconPosition={props.expandIconPosition}
-        items={collapsibleItems}
-        onChange={handleCollapsibleChange}
-      />
+      {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions */}
+      <div onClick={(e) => e.stopPropagation()}>
+        <Collapse
+          accordion={values?.isAccordion}
+          activeKey={activeValue}
+          expandIcon={expandIcon}
+          expandIconPosition={props.expandIconPosition}
+          items={collapsibleItems}
+          onChange={handleCollapsibleChange}
+        />
+      </div>
     </ConfigProvider>
   );
 };