+++ packages/runtime/src/widgets/ToggleButton.tsx
@@ -32,7 +32,7 @@ interface ToggleButtonPropsStyles {
   selectedBorderWidth?: number;
 }
 
-type ToggleButtonProps = {
+export type ToggleButtonProps = {
   value: string;
   items:
     | string[]
@@ -83,19 +83,19 @@ export const ToggleButton: React.FC<ToggleButtonProps> = ({
 
   return (
     <MUIToggleButtonGroup
+      className={values?.styles?.names}
       exclusive
       onChange={onChangeCallback}
       ref={rootRef}
-      className={values?.styles?.names}
       sx={{
-        display: "flex",
-        flexDirection: "row",
-        flexWrap: "wrap",
+        display: "grid",
+        gridTemplateColumns: `repeat(${structuredItems.length},1fr)`,
         rowGap: `${styles?.runSpacing ?? 0}px`,
         margin: `${styles?.margin ?? 0}`,
         ...(values?.styles?.visible === false
           ? { display: "none" }
           : undefined),
+        ...values?.styles,
       }}
       value={values?.value}
     >
@@ -139,7 +139,9 @@ export const ToggleButton: React.FC<ToggleButtonProps> = ({
           }}
           value={item.value}
         >
-          {isString(item.icon?.name) && <Icon {...item.icon!} />}
+          {item.icon && isString(item.icon.name) ? (
+            <Icon {...item.icon} />
+          ) : null}
           &nbsp;
           <Text text={item.label} />
         </MUIToggleButton>