+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ensembleui/react-starter
 
+## 0.3.17
+
+### Patch Changes
+
+- Updated dependencies [36e127e]
+- Updated dependencies [602ac3e]
+  - @ensembleui/react-runtime@0.1.28
+
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [70e700d]
+  - @ensembleui/react-runtime@0.1.27
+
+## 0.3.15
+
+### Patch Changes
+
+- Updated dependencies [31583e0]
+- Updated dependencies [d55d0d2]
+- Updated dependencies [e12ddd6]
+  - @ensembleui/react-runtime@0.1.26
+  - @ensembleui/react-framework@0.1.20
+
 ## 0.3.14
 
 ### Patch Changes